requirements.txt
anyascii==0.3.2
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==4.1.0
django-environ==0.9.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
draftjs-exporter==2.1.7
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
environ==1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
l18n==2021.3
openpyxl==3.1.2
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
telepath==0.3
uritemplate==4.1.1
urllib3==1.26.15
wagtail==4.2.2
webencodings==0.5.1
Willow==1.4.1
yandex-geocoder==3.0.1
